+++ GluonicLQCD/observables/tools/observablestorer.cpp
@@ -5,7 +5,7 @@
 #include <cmath>
 #include <mpi.h>
 
-ObservableStorer::ObservableStorer(int NSize)
+ObservableStorer::ObservableStorer(unsigned long int NSize)
 {
     /*
      * Initialization of a observable storage container.
@@ -16,7 +16,7 @@ ObservableStorer::ObservableStorer(int NSize)
     // Initializes arrays for storing the observables
     m_observables = new double[m_NObs];
     m_observablesSquared = new double[m_NObs];
-    for (int iObs = 0; iObs < m_NObs; iObs++) m_observables[iObs] = 0;
+    for (unsigned long int iObs = 0; iObs < m_NObs; iObs++) m_observables[iObs] = 0;
 }
 
 ObservableStorer::~ObservableStorer()
@@ -33,7 +33,7 @@ void ObservableStorer::gatherResults()
     // Temporary buffer for summing the observables
     double tempBuffer[m_NObs];
 
-    for (int iObs = 0; iObs < m_NObs; iObs++) {
+    for (unsigned long int iObs = 0; iObs < m_NObs; iObs++) {
         tempBuffer[iObs] = 0;
     }
 
@@ -41,14 +41,14 @@ void ObservableStorer::gatherResults()
     MPI_Reduce(m_observables, tempBuffer, m_NObs, MPI_DOUBLE, MPI_SUM, 0, Parallel::ParallelParameters::ACTIVE_COMM);
 
     // Retrieving observable from temporary buffer
-    for (int iBuffer = 0; iBuffer < m_NObs; iBuffer++) {
+    for (unsigned long int iBuffer = 0; iBuffer < m_NObs; iBuffer++) {
         m_observables[iBuffer] = tempBuffer[iBuffer];
     }
 
     // Normalizing by the number of processors if specified
     if (m_normalizeObservableByProcessor) {
         double numprocs = double(Parallel::Communicator::getNumProc());
-        for (int i = 0; i < m_NObs; i++) {
+        for (unsigned long int i = 0; i < m_NObs; i++) {
             m_observables[i] /= double(numprocs);
         }
     }
@@ -60,11 +60,11 @@ void ObservableStorer::runStatistics()
      * Performs the statistics on the observables.
      * Best used in conjecture with ObservableStorer::gatherResults().
      */
-    for (int i = 0; i < m_NObs; i++) {
+    for (unsigned long int i = 0; i < m_NObs; i++) {
         m_observablesSquared[i] = m_observables[i]*m_observables[i];
     }
     // Gets average of the observable
-    for (int i = 0; i < m_NObs; i++)
+    for (unsigned long int i = 0; i < m_NObs; i++)
     {
         m_averagedObservable += m_observables[i];
         m_averagedObservableSquared += m_observablesSquared[i];
@@ -98,7 +98,7 @@ void ObservableStorer::writeObservableToFile(double acceptanceRatio)
                                m_stdObservable, m_observables, m_NObs, m_observableName);
 }
 
-void ObservableStorer::writeFlowObservableToFile(int configNumber)
+void ObservableStorer::writeFlowObservableToFile(unsigned long configNumber)
 {
     /*
      * Writes a flow observable to file
@@ -111,7 +111,7 @@ void ObservableStorer::reset()
     /*
      * Sets all the observables to zero. A must-have when flowing.
      */
-    for (int i = 0; i < m_NObs; i++)
+    for (unsigned long int i = 0; i < m_NObs; i++)
     {
         m_observables[i] = 0;
         m_observablesSquared[i] = 0;
